Fixes bug 1090248 Split out PostgreSQLAlchemyManager
